+++ src/server/admin/js/index.js
@@ -3,7 +3,9 @@ const axios = require('axios')
 /**
  * Initialize the admin dashboard.
  *
- * @since unreleased
+ * @since 1.5.0
+ * @since unreleased Save pledges data to sessionStorage and reuse for
+ * subsequent page loads.
  */
 function init () {
   // Get pledges data from sessionStorage.
@@ -21,8 +23,8 @@ function init () {
       .catch(error => {
         // Log the error and display it in the UI.
         console.error(error)
-        const loading = document.querySelector('[data-admin="loading"]')
-        loading.textContent = `Error: ${error}`
+        showMessage('error', error)
+
       })
   } else {
     // Render the admin dashboard with pledges data from sessionStorage.
@@ -31,16 +33,47 @@ function init () {
 }
 
 /**
- * Render admin dashboard.
+ * Show UI message.
  *
  * @since unreleased
+ *
+ * @param {string} state   The state.
+ * @param {string} message The message.
+ */
+function showMessage (state, message) {
+  // Set state and remove any child elements.
+  const container = document.querySelector('[data-admin="message"]')
+  container.setAttribute('data-state', state)
+  container.innerHTML = ''
+
+  // Add message.
+  const paragraph = document.createElement('p')
+  paragraph.textContent = message
+  container.appendChild(paragraph)
+}
+
+/**
+ * Clear message.
+ * 
+ * @since unreleased
+ */
+function clearMessage () {
+  const container = document.querySelector('[data-admin="message"]')
+  // Update state and remove any child elements.
+  container.setAttribute('data-state', 'inactive')
+  container.innerHTML = ''
+}
+
+/**
+ * Render admin dashboard.
+ *
+ * @since 1.5.0
  * 
  * @param {array} pledges Array of pledge objects.
  */
 function renderDashboard (pledges) {
   // Remove loading message.
-  const loading = document.querySelector('[data-admin="loading"]')
-  loading.parentElement.removeChild(loading)
+  clearMessage()
 
   // Render the dashboard.
   const dashboard = document.querySelector('[data-admin="dashboard"]')
@@ -50,12 +83,24 @@ function renderDashboard (pledges) {
 /**
  * Render admin dashboard markup.
  *
- * @since unreleased
+ * @since 1.5.0
  * 
  * @param {array} pledges Array of pledge objects.
  */
 function getDashboardMarkup (pledges) {
-  // Build the data.
+  /**
+   * Filter pledges data by brand.
+   *
+   * @since 1.5.0
+   *
+   * @param  {string}  string  String to filter pledge brand name by.
+   * @return {array}           Filtered array of pledge objects.
+   */
+  function getPledgesBrand(string) {
+    return pledges.filter(pledge => pledge.brand.includes(string))
+  }
+
+  // Build brands pledges data.
   const brands = {
     bhgre: { pledges: getPledgesBrand('Better Homes') },
     c21: { pledges: getPledgesBrand('Century') },
@@ -69,11 +114,6 @@ function getDashboardMarkup (pledges) {
     total: { pledges: pledges }
   }
 
-  function getPledgesBrand(string) {
-    return pledges.filter(pledge => pledge.brand.includes(string))
-  }
-
-  // Build brands data.
   Object.keys(brands).forEach(brand => {
     // Add brand name.
     brands[brand].name = brands[brand].pledges[0].brand
@@ -87,7 +127,7 @@ function getDashboardMarkup (pledges) {
         .length
   })
 
-  // Custommize brand names.
+  // Update brand names.
   brands.bhgre.name = 'BHGRE'
   brands.c21.name = 'Century 21'
   brands.cb.name = 'Coldwell Banker'